+++ api/server.go
@@ -26,7 +26,7 @@ var router *routing.Router
 
 func main() {
 	config.LoadConfigs()
-
+	
 	router = routing.New()
 	router.Get("/reload", authentication.AuthorizationMiddleware, ReloadServices)
 	router.Get("/invalidate-cache", authentication.AuthorizationMiddleware, InvalidateCache)
